requirements.txt
aiogram==3.12.0
django==5.1
python-decouple==3.8
redis==5.0.8
requests==2.32.3